+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### v0.3.11
+#### Added
+- Created a Troubleshooting page, comprising Diagnostics and Self-tests.
+#### Updated
+- Made the buttons, icons, and links on the catalog page visually consistent with the rest of the interface.
+#### Fixed
+- Fixed an issue with the alignment of delete buttons in the Input List component. 
+
 ### v0.3.10
 #### Updated
 - Updating the scss files so they better match one-to-one with React component files.